refactor(analyzer): pass 2 — move test-only ctors, clean dead code
/simplify pass 2. Three small wins:

1. Move GetBaseMetadata + NewGalaAnalyzerWithBase to a _test.go file
   inside the analyzer package. Both had zero production callers
   (verified `grep -rn ... --include='*.go' . | grep -v _test.go`)
   and were documented as test-only. Putting them in
   test_exports_test.go keeps the existing analyzer_test package's
   call site working while removing them from the production API
   surface — and from the production binary.

2. Simplify computeExtractIndices: drop the unused method receiver
   (logic operates only on parameters, so a free function is
   simpler), inline the redundant normalizedContainerParam local,
   and delete the empty if-block that contained nothing but a
   comment about why it was empty.

3. pruneStaleCaches's defer-recover now logs the recovered value to
   stderr. Best-effort cache cleanup should still leave a
   breadcrumb when filesystem operations panic — silently absorbing
   permission errors / OS-level corruption masks real problems.

Out of scope (deferred):
  - Functional-options refactor of the four analyzer constructors
    (NewGalaAnalyzer / *WithPackageFiles / *ForLSP)
  - ImportManager naming consistency sweep
  - Narrowing the Analyzer interface for LSP

All 331 tests pass.
